Move form logger widget to its own package
